Updates in integrating Blocks and Files
